+++ CHANGELOG.md
@@ -6,8 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Fixed
-- `stroke-miterlimit` will be parsed as `Number` and not as a `Length` now.
+- `stroke-miterlimit` will be parsed as a `Number` and not as a `Length` now.
 - `number` attribute type parsing.
+- `offset` attribute parsing.
 
 ## [0.15.0] - 2018-12-13
 ### Added